typescript-sdk/integrations/crewai/python/pyproject.toml
[tool.poetry]
name = "agui_crewai"
version = "0.1.0"
description = ""
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.14,>=3.10"
ag-ui-protocol = "==0.1.5"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
crewai = "^0.130.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "agui_crewai.dojo:main"